+++ src/ai/ai_war.cpp
@@ -1,5 +1,8 @@
 #include "ai_war.hpp"
 #include "ai_campaign_values.hpp"
+#include "prng.hpp"
+#include "demographics.hpp"
+#include "triggers.hpp"
 
 namespace ai {
